requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
click>=8.0.0
rich>=13.0.0
loguru>=0.7.0

# AWS dependencies
boto3>=1.34.0
botocore>=1.34.0

# RAG and AI dependencies
openai>=1.0.0
tiktoken>=0.5.0
vecclean>=0.1.0

# Document processing
pypdf2>=3.0.0
python-docx>=1.1.0
markdown>=3.5.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Web framework (optional, for future API)
fastapi>=0.100.0
uvicorn>=0.23.0

# Testing (dev dependencies)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0